+++ lucid/utils/cache/__init__.py
@@ -11,9 +11,11 @@ models use :class:`EncoderDecoderCache`.  Canonical paths:
 from lucid.utils.cache._base import Cache
 from lucid.utils.cache._dynamic import DynamicCache
 from lucid.utils.cache._encoder_decoder import EncoderDecoderCache
+from lucid.utils.cache._static import StaticCache
 
 __all__ = [
     "Cache",
     "DynamicCache",
     "EncoderDecoderCache",
+    "StaticCache",
 ]